+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/analytics/Type.kt
@@ -42,5 +42,6 @@ internal enum class Actype(val value: String) {
     JSON_INFO("mini_app_send_json_to_host_app"),
     CLOSE_MINIAPP("mini_app_close"),
     GET_HOST_APP_THEME_COLORS("mini_app_get_host_app_theme_colors"),
+    GET_IS_DARK_MODE("mini_app_get_is_dark_mode"),
     DEFAULT("")
 }